requirements.txt
python-dotenv
python-telegram-bot
openai
google-cloud-storage
pydub
nest_asyncio